requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
decorator==5.1.1
distlib==0.3.8
Django==4.2.17
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.0.1
Faker==28.4.1
fastapi==0.115.7
fastapi-cli==0.0.7
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipdb==0.13.13
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
orjson==3.10.15
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipenv==2024.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
psycopg==3.2.4
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.44.0
tabulate==0.9.0
tomli==2.0.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.33.0
uvloop==0.21.0
virtualenv==20.26.6
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.4
zipp==3.20.2
